dev-requirements.txt
setuptools>=0.6c9
simplejson
Mock>=0.5.0
paver-templates>=0.1.0b3
Nose>=0.10.4
GitPython>=0.1.6
virtualenv>=1.3.3
